внес поправки, чтобы сервер заработал
