+++ orders/models.py
@@ -1,16 +1,22 @@
 from django.db import models
+from django.contrib.auth import get_user_model
+
+from products.models import Product
+
+User = get_user_model()
 
 
 class Orders(models.Model):
     article_number = models.CharField(max_length=50)
     customer = models.ForeignKey(
-        "User",
+        User,
         on_delete=models.CASCADE,
         related_name="orders",
     )
     address = models.ForeignKey(
         "DeliveryAddress",
         on_delete=models.SET_NULL,
+        null=True
     )
     created_at = models.DateTimeField(auto_now_add=True)
     price_total = models.FloatField()
@@ -22,10 +28,12 @@ class OrderProducts(models.Model):
         Orders,
         on_delete=models.SET_NULL,
         related_name="orderProducts",
+        null=True
     )
     product_id = models.ForeignKey(
-        "Products",
+        Product,
         on_delete=models.SET_NULL,
+        null=True
     )
     amount = models.FloatField()
     purchase_price = models.FloatField()
@@ -33,7 +41,7 @@ class OrderProducts(models.Model):
 
 class DeliveryAddress(models.Model):
     owner = models.ForeignKey(
-        "User",
+        User,
         on_delete=models.CASCADE,
         related_name="deliveryAddress",
     )
